requirements.txt
matplotlib>=3.7
numpy>=1.23
scipy>=1.10
pyyaml>=6.0
